Move autoconfiguration to new namespace
